requirements.txt
Flask==2.3.3
openai==1.55.3
httpx==0.27.2 
pdf2image==1.16.3
Pillow==10.1.0
PyPDF2==3.0.1
python-dotenv==1.0.0
Werkzeug==2.3.7
requests==2.31.0